refactor(frontend): refactor the folder structure and rename the footer component
