+++ apps/iaimc-backend/src/app/configurations/database.config.ts
@@ -24,10 +24,10 @@ export default (): DatabaseConfig => ({
     url: process.env.POSTGRES_DATABASE_URL,
     synchronize: process.env.NODE_ENV === 'development',
     ssl: false,
-    extra: {
-      ssl: {
-        rejectUnauthorized: false,
-      },
-    },
+    // extra: {
+    //   ssl: {
+    //     rejectUnauthorized: false,
+    //   },
+    // },
   } satisfies TypeOrmModuleOptions,
 });
