+++ apps/iaimc-frontend/public/assets/images/thank-you.png
