+++ apps/iaimc-frontend/src/app/app.effects.ts
@@ -1,3 +1,3 @@
-import { FooterEffects } from 'libs/ui-components/src/footer/store/footer.effects';
+import { FooterEffects } from '../../../../libs/ui-components/src/model/components/main-footer/store/footer.effects';
 
 export const appEffects = [FooterEffects];
