+++ apps/iaimc-frontend/src/app/app.reducer.ts
@@ -1,7 +1,5 @@
-import {
-  FOOTER_FEATURE_KEY,
-  footerReducer,
-} from 'libs/ui-components/src/footer/store/reducers/footer.reducer';
+import { FOOTER_FEATURE_KEY, footerReducer } from "../../../../libs/ui-components/src/model/components/main-footer/store/footer.reducer";
+
 
 export const appReducer = {
   [FOOTER_FEATURE_KEY]: footerReducer,
