+++ apps/iaimc-frontend/src/app/customer/library/library.component.html
@@ -1,3 +1,4 @@
 <app-stage />
 <app-model-list />
+<app-main-footer/>
 <app-footer />
