+++ apps/iaimc-frontend/src/styles/stylestheme.scss
@@ -1,8 +1,20 @@
+
+
 /* Note: Color palettes are generated from primary: #7F00FF */
 html {
   /* CUSTOM VARIABLES */
   --toolbar-height: 50px;
   --navbar-height: 64px;
+  --toolbar-height: 50px;
+  --navbar-height: 64px;
+  --font-size-footer: 0.875rem;
+  --footer-links-color: #606060;
+  --footer-main-background: #1e1925;
+  --footer-border-color: #484848;
+  --subscription-background: #262626;
+  --subscription-primary-color:#A1A1A1;
+  --footer-icon-background-color:#606060;
+
 
   /* COLOR SYSTEM VARIABLES */
   color-scheme: light;
