+++ libs/ui-components/src/index.ts
@@ -1,6 +1,6 @@
-import * as feedBackI18n from './footer/components/thank-you-feedback/feedback.i18n.json';
-import * as footerI18n from './footer/footer.i18n.json';
-import * as landingPageFooterI18n from './landing-page/landing-page-footer/landing-page.i18n.json';
+import * as feedBackI18n from './model/components/main-footer/components/thank-you-feedback/feedback.i18n.json';
+import * as footerI18n from './model/components/main-footer/footer.i18n.json';
+import * as footerSubscriptionI18n from './landing-page/components/footer/footer.i18n.json';
 import * as langSwticherI18n from './language-switcher/lang-switcher.i18n.json';
 import * as modelListI18n from './model/components/model-list/model-list.i18n.json';
 import * as modelUploadI18n from './model/components/model-upload/model-upload.i18n.json';
@@ -12,9 +12,8 @@ import * as toolbarI18n from './toolbar/toolbar.i18n.json';
 export { toolbarI18n };
 export { stageI18n };
 export { langSwticherI18n };
-export * from './footer/footer.component';
 export { feedBackI18n };
-export { footerI18n, landingPageFooterI18n };
+export { footerI18n, footerSubscriptionI18n };
 export { navBarI18n };
 export { modelListI18n, modelUploadI18n };
 
@@ -24,6 +23,6 @@ export * from './model/components/model-upload/model-upload.component';
 export * from './stage/stage.component';
 export * from './language-switcher/constants';
 export * from './language-switcher/language-switcher.component';
-export * from './landing-page/landing-page-footer/landing-page-footer.component';
+export * from './model/components/main-footer/main-footer.component';
 export * from './navbar/navbar.component';
 export * from './toolbar/toolbar.component';
