+++ libs/ui-components/src/landing-page/components/footer/footer.i18n.json
@@ -1,5 +1,5 @@
 {
-  "landing-page-footer": {
+  "footer": {
     "LOGO_TEXT": {
       "ar": "مصمم نموذج الحقن",
       "de": "Spritzgussmodell-Designer",
