+++ libs/ui-components/src/model/components/main-footer/components/thank-you-feedback/feedback.i18n.json
