+++ libs/ui-components/src/model/components/main-footer/components/thank-you-feedback/thank-you-feedback.component.html
@@ -38,7 +38,7 @@
     <div class="success-container">
       <h2>{{ 'feedback.THANKS' | transloco }}</h2>
       <div class="feedback-image">
-        <img src="thank-you.png" alt="thank you" class="success-image" />
+        <img src="assets/images/thank-you.png" alt="thank you" class="success-image" />
       </div>
     </div>
   }
