+++ libs/ui-components/src/model/components/main-footer/components/thank-you-feedback/thank-you-feedback.component.scss
