+++ libs/ui-components/src/model/components/main-footer/footer.i18n.json
