+++ libs/ui-components/src/model/components/main-footer/footer.interface.ts
