+++ libs/ui-components/src/model/components/main-footer/main-footer.component.html
@@ -7,8 +7,8 @@
           <div class="contact__tel">
             <p>{{ 'footer.APP' | transloco }}</p>
             <a [href]="'tel:+4922039649145'">{{
-              'footer.PHONE' | transloco
-            }}</a>
+                'footer.PHONE' | transloco
+              }}</a>
           </div>
           <button mat-button class="custom-button">
             {{ 'footer.WRITE_A_MESSAGE' | transloco }}
