+++ libs/ui-components/src/model/components/main-footer/main-footer.component.scss
@@ -17,10 +17,10 @@ mat-toolbar {
   height: auto;
 
   @include mat.toolbar-overrides(
-    (
-      container-background-color: var(--mat-sys-on-secondary-fixed),
-      container-text-color: var(--mat-sys-primary-container),
-    )
+      (
+        container-background-color: var(--mat-sys-on-secondary-fixed),
+        container-text-color: var(--mat-sys-primary-container),
+      )
   );
 
   footer {
@@ -85,9 +85,9 @@ mat-toolbar {
         }
 
         @include mat.progress-spinner-overrides(
-          (
-            active-indicator-color: var(--mat-sys-primary),
-          )
+            (
+              active-indicator-color: var(--mat-sys-primary),
+            )
         );
 
         .rating-stars {
