+++ libs/ui-components/src/model/components/main-footer/main-footer.component.ts
@@ -1,18 +1,5 @@
-import { CommonModule } from '@angular/common';
-import {
-  ChangeDetectionStrategy,
-  Component,
-  inject,
-  OnDestroy,
-  OnInit,
-  signal,
-} from '@angular/core';
-import {
-  FormBuilder,
-  FormsModule,
-  ReactiveFormsModule,
-  Validators,
-} from '@angular/forms';
+import { ChangeDetectionStrategy, Component, inject, OnDestroy, OnInit, signal } from '@angular/core';
+import { FormBuilder, FormsModule, ReactiveFormsModule, Validators } from '@angular/forms';
 import { MatButtonModule } from '@angular/material/button';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatIconModule } from '@angular/material/icon';
@@ -22,21 +9,17 @@ import { MatToolbarModule } from '@angular/material/toolbar';
 import { TranslocoPipe } from '@jsverse/transloco';
 import { Store } from '@ngrx/store';
 import { Subject, takeUntil } from 'rxjs';
-import { formField } from '../utilities/helper-function';
-import { atLeastOneFieldValidator } from '../validators/custom-validators/custom.validator';
+import { formField } from '../../../utilities/helper-function';
+import { atLeastOneFieldValidator } from '../../../validators/custom.validator';
 import { ThankYouFeedbackComponent } from './components/thank-you-feedback/thank-you-feedback.component';
 import { FeedbackRequest } from './footer.interface';
 import { FooterService } from './service/footer.service';
 import { FooterActions } from './store/footer.actions';
-import {
-  selectFeedbackLoading,
-  selectIsFeedbackSubmitted,
-} from './store/selectors/footer.selectors';
+import { selectFeedbackLoading, selectIsFeedbackSubmitted } from './store/footer.selectors';
 
 @Component({
-  selector: 'app-footer',
+  selector: 'app-main-footer',
   imports: [
-    CommonModule,
     MatToolbarModule,
     MatButtonModule,
     MatIconModule,
@@ -48,11 +31,11 @@ import {
     ThankYouFeedbackComponent,
     TranslocoPipe,
   ],
-  templateUrl: './footer.component.html',
-  styleUrl: './footer.component.scss',
+  templateUrl: './main-footer.component.html',
+  styleUrl: './main-footer.component.scss',
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export class FooterComponent implements OnInit, OnDestroy {
+export class MainFooterComponent implements OnInit, OnDestroy {
   private readonly fb = inject(FormBuilder);
   private readonly store = inject(Store);
   private readonly footerService = inject(FooterService);
