+++ libs/ui-components/src/model/components/main-footer/service/footer.service.spec.ts
@@ -7,7 +7,8 @@ import { FooterService } from './footer.service';
 import { provideHttpClient } from '@angular/common/http';
 import { MockStore, provideMockStore } from '@ngrx/store/testing';
 import { FeedbackRequest, FeedbackResponse } from '../footer.interface';
-import { selectFeedbackId } from '../store/selectors/footer.selectors';
+import { selectFeedbackId } from '../store/footer.selectors';
+
 
 describe('FooterService', () => {
   let service: FooterService;
