+++ libs/ui-components/src/model/components/main-footer/store/footer.actions.ts
