+++ libs/ui-components/src/model/components/main-footer/store/footer.reducer.spec.ts
@@ -1,6 +1,7 @@
-import { FooterActions } from '../footer.actions';
-import { initialFooterState } from '../footer.state';
+import { FooterActions } from './footer.actions';
 import { footerReducer } from './footer.reducer';
+import { initialFooterState } from './footer.state';
+
 
 describe('Footer Reducer', () => {
   it('should start feedback submission by setting isFeedbackLoading to true', () => {
