+++ libs/ui-components/src/validators/text-only.validator.ts
